+++ tests/test_write_gate.py
@@ -1,6 +1,6 @@
 import unittest
 
-from lmlib.write_gate import WriteGate
+from openlmlib.write_gate import WriteGate
 
 
 class FixedEmbedder: